Move public ModelReaderWriter API changes in
