+++ sdk/core/System.Net.ClientModel/src/ModelReaderWriter/ModelReaderWriter.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Net.ClientModel.Core;
-using System.Net.ClientModel.Internal;
 
 namespace System.Net.ClientModel
 {
@@ -23,11 +22,21 @@ namespace System.Net.ClientModel
         /// <exception cref="ArgumentNullException">If <paramref name="model"/> is null.</exception>
         public static BinaryData Write<T>(T model, ModelReaderWriterOptions? options = default) where T : IModel<T>
         {
-            ClientUtilities.AssertNotNull(model, nameof(model));
+            if (model is null)
+                throw new ArgumentNullException(nameof(model));
 
             options ??= ModelReaderWriterOptions.DefaultWireOptions;
 
-            return model.Write(options);
+            if (model is IJsonModel<T> jsonModel &&
+                (options.Format == ModelReaderWriterFormat.Json || (options.Format == ModelReaderWriterFormat.Wire && model.GetWireFormat(options) == ModelReaderWriterFormat.Json)))
+            {
+                using ModelWriter<T> writer = new ModelWriter<T>(jsonModel, options);
+                return writer.ToBinaryData();
+            }
+            else
+            {
+                return model.Write(options);
+            }
         }
 
         /// <summary>
@@ -54,7 +63,8 @@ namespace System.Net.ClientModel
         /// <exception cref="ArgumentNullException">If <paramref name="model"/> is null.</exception>
         public static BinaryData Write(object model, ModelReaderWriterOptions? options = default)
         {
-            ClientUtilities.AssertNotNull(model, nameof(model));
+            if (model is null)
+                throw new ArgumentNullException(nameof(model));
 
             options ??= ModelReaderWriterOptions.DefaultWireOptions;
 
@@ -64,7 +74,16 @@ namespace System.Net.ClientModel
                 throw new InvalidOperationException($"{model.GetType().Name} does not implement {nameof(IModel<object>)}");
             }
 
-            return iModel.Write(options);
+            if (iModel is IJsonModel<object> jsonObject &&
+                (options.Format == ModelReaderWriterFormat.Json || (options.Format == ModelReaderWriterFormat.Wire && iModel.GetWireFormat(options) == ModelReaderWriterFormat.Json)))
+            {
+                using ModelWriter writer = new ModelWriter(jsonObject, options);
+                return writer.ToBinaryData();
+            }
+            else
+            {
+                return iModel.Write(options);
+            }
         }
 
         /// <summary>
@@ -90,7 +109,8 @@ namespace System.Net.ClientModel
         /// <exception cref="ArgumentNullException">If <paramref name="data"/> is null.</exception>
         public static T? Read<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] T>(BinaryData data, ModelReaderWriterOptions? options = default) where T : IModel<T>
         {
-            ClientUtilities.AssertNotNull(data, nameof(data));
+            if (data is null)
+                throw new ArgumentNullException(nameof(data));
 
             options ??= ModelReaderWriterOptions.DefaultWireOptions;
 
@@ -123,8 +143,10 @@ namespace System.Net.ClientModel
         /// <exception cref="ArgumentNullException">If <paramref name="data"/> or <paramref name="returnType"/> are null.</exception>
         public static object? Read(BinaryData data, [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type returnType, ModelReaderWriterOptions? options = default)
         {
-            ClientUtilities.AssertNotNull(data, nameof(data));
-            ClientUtilities.AssertNotNull(returnType, nameof(returnType));
+            if (data is null)
+                throw new ArgumentNullException(nameof(data));
+            if (returnType is null)
+                throw new ArgumentNullException(nameof(returnType));
 
             options ??= ModelReaderWriterOptions.DefaultWireOptions;
 
@@ -144,22 +166,6 @@ namespace System.Net.ClientModel
         /// <exception cref="ArgumentNullException">If <paramref name="data"/> or <paramref name="returnType"/> are null.</exception>
         public static object? Read(BinaryData data, [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type returnType, ModelReaderWriterFormat format)
             => Read(data, returnType, ModelReaderWriterOptions.GetOptions(format));
-
-        /// <summary>
-        /// Converts the value of a model into a <see cref="BinaryData"/>.
-        /// </summary>
-        /// <param name="model">The model to convert.</param>
-        /// <param name="options">The <see cref="ModelReaderWriterOptions"/> to use.</param>
-        /// <returns>A <see cref="BinaryData"/> representation of the model in the <see cref="ModelReaderWriterFormat"/> specified by the <paramref name="options"/>.</returns>
-        /// <exception cref="ArgumentNullException">If <paramref name="model"/> or <paramref name="options"/> are null.</exception>
-        public static BinaryData WriteCore(IJsonModel<object> model, ModelReaderWriterOptions options)
-        {
-            ClientUtilities.AssertNotNull(model, nameof(model));
-            ClientUtilities.AssertNotNull(options, nameof(options));
-
-            using ModelWriter writer = new ModelWriter(model, options);
-            return writer.ToBinaryData();
-        }
 
         private static IModel<object> GetInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type returnType)
         {
